+++ examples/abc.rs
@@ -7,28 +7,28 @@ fn main() {
 
     let capacity = num_poppers * num_ticks;
 
-    let queue = ConcurrentQueue::new();
-    let collected = ConcurrentBag::new();
+    // let queue = ConcurrentQueue::new();
+    // let collected = ConcurrentBag::new();
 
-    for i in 0..capacity {
-        queue.push(i);
-    }
+    // for i in 0..capacity {
+    //     queue.push(i);
+    // }
 
-    std::thread::scope(|s| {
-        for _ in 0..num_poppers {
-            s.spawn(|| {
-                for _ in 0..(num_ticks + 10) {
-                    let popped = queue.pop();
-                    // println!("{popped:?}");
-                    if let Some(value) = popped {
-                        collected.push(value);
-                    }
-                }
-            });
-        }
-    });
+    // std::thread::scope(|s| {
+    //     for _ in 0..num_poppers {
+    //         s.spawn(|| {
+    //             for _ in 0..(num_ticks + 10) {
+    //                 let popped = queue.pop();
+    //                 // println!("{popped:?}");
+    //                 if let Some(value) = popped {
+    //                     collected.push(value);
+    //                 }
+    //             }
+    //         });
+    //     }
+    // });
 
-    let mut collected = collected.into_inner();
-    collected.sort();
-    assert_eq!(collected, (0..collected.len()).collect::<Vec<_>>());
+    // let mut collected = collected.into_inner();
+    // collected.sort();
+    // assert_eq!(collected, (0..collected.len()).collect::<Vec<_>>());
 }
