+++ src/arch2/tests/mod.rs
@@ -1,0 +1,6 @@
+mod extend;
+mod pop;
+mod pull;
+mod pull_extend;
+mod push;
+mod push_pop;
