--- src/tests/mod.rs
@@ -1,6 +1,0 @@
-mod extend;
-mod pop;
-mod pull;
-mod pull_extend;
-mod push;
-mod push_pop;
